api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
db-dtypes==1.4.3
fastapi==0.115.12
google-api-core==2.24.2
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
numpy==2.2.6
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pandas-gbq==0.29.0
pluggy==1.6.0
polars==1.29.0
proto-plus==1.26.1
protobuf==5.29.4
pyaml==25.1.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.5
pydantic-core==2.33.2
pydata-google-auth==1.9.1
pysocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.32.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
trio==0.30.0
trio-websocket==0.12.2
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websocket-client==1.8.0
wsproto==1.2.0